+++ src/main/java/com/kard/api/resources/users/AsyncRawUsersClient.java
@@ -22,6 +22,7 @@ import com.kard.api.resources.users.types.CreateUsersMultiStatusResponse;
 import com.kard.api.resources.users.types.CreateUsersObject;
 import com.kard.api.resources.users.types.DeleteUserResponseObject;
 import com.kard.api.resources.users.types.UpdateUserObject;
+import com.kard.api.resources.users.types.UserResponseObject;
 import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
 import okhttp3.Call;
@@ -152,7 +153,7 @@ public class AsyncRawUsersClient {
      * Call this endpoint to update the details on a specified user.<br/>
      * <p><b>Required scopes:</b> <code>user:update</code></p>
      */
-    public CompletableFuture<KardApiHttpResponse<UpdateUserObject>> update(
+    public CompletableFuture<KardApiHttpResponse<UserResponseObject>> update(
             String organizationId, String userId, UpdateUserObject request) {
         return update(organizationId, userId, request, null);
     }
@@ -161,7 +162,7 @@ public class AsyncRawUsersClient {
      * Call this endpoint to update the details on a specified user.<br/>
      * <p><b>Required scopes:</b> <code>user:update</code></p>
      */
-    public CompletableFuture<KardApiHttpResponse<UpdateUserObject>> update(
+    public CompletableFuture<KardApiHttpResponse<UserResponseObject>> update(
             String organizationId, String userId, UpdateUserObject request, RequestOptions requestOptions) {
         HttpUrl.Builder httpUrl = HttpUrl.parse(this.clientOptions.environment().getUrl())
                 .newBuilder()
@@ -192,7 +193,7 @@ public class AsyncRawUsersClient {
         if (requestOptions != null && requestOptions.getTimeout().isPresent()) {
             client = clientOptions.httpClientWithTimeout(requestOptions);
         }
-        CompletableFuture<KardApiHttpResponse<UpdateUserObject>> future = new CompletableFuture<>();
+        CompletableFuture<KardApiHttpResponse<UserResponseObject>> future = new CompletableFuture<>();
         client.newCall(okhttpRequest).enqueue(new Callback() {
             @Override
             public void onResponse(@NotNull Call call, @NotNull Response response) throws IOException {
@@ -200,7 +201,7 @@ public class AsyncRawUsersClient {
                     String responseBodyString = responseBody != null ? responseBody.string() : "{}";
                     if (response.isSuccessful()) {
                         future.complete(new KardApiHttpResponse<>(
-                                ObjectMappers.JSON_MAPPER.readValue(responseBodyString, UpdateUserObject.class),
+                                ObjectMappers.JSON_MAPPER.readValue(responseBodyString, UserResponseObject.class),
                                 response));
                         return;
                     }
@@ -343,7 +344,7 @@ public class AsyncRawUsersClient {
      * <br/>
      * <b>Required scopes:</b>  <code>user:read</code>
      */
-    public CompletableFuture<KardApiHttpResponse<UpdateUserObject>> get(String organizationId, String userId) {
+    public CompletableFuture<KardApiHttpResponse<UserResponseObject>> get(String organizationId, String userId) {
         return get(organizationId, userId, null);
     }
 
@@ -352,7 +353,7 @@ public class AsyncRawUsersClient {
      * <br/>
      * <b>Required scopes:</b>  <code>user:read</code>
      */
-    public CompletableFuture<KardApiHttpResponse<UpdateUserObject>> get(
+    public CompletableFuture<KardApiHttpResponse<UserResponseObject>> get(
             String organizationId, String userId, RequestOptions requestOptions) {
         HttpUrl.Builder httpUrl = HttpUrl.parse(this.clientOptions.environment().getUrl())
                 .newBuilder()
@@ -375,7 +376,7 @@ public class AsyncRawUsersClient {
         if (requestOptions != null && requestOptions.getTimeout().isPresent()) {
             client = clientOptions.httpClientWithTimeout(requestOptions);
         }
-        CompletableFuture<KardApiHttpResponse<UpdateUserObject>> future = new CompletableFuture<>();
+        CompletableFuture<KardApiHttpResponse<UserResponseObject>> future = new CompletableFuture<>();
         client.newCall(okhttpRequest).enqueue(new Callback() {
             @Override
             public void onResponse(@NotNull Call call, @NotNull Response response) throws IOException {
@@ -383,7 +384,7 @@ public class AsyncRawUsersClient {
                     String responseBodyString = responseBody != null ? responseBody.string() : "{}";
                     if (response.isSuccessful()) {
                         future.complete(new KardApiHttpResponse<>(
-                                ObjectMappers.JSON_MAPPER.readValue(responseBodyString, UpdateUserObject.class),
+                                ObjectMappers.JSON_MAPPER.readValue(responseBodyString, UserResponseObject.class),
                                 response));
                         return;
                     }